requirements.txt
flask
aiogram
python-dotenv
gigachat
asyncio